Reorganize API docs (#24)
## Changes

### Adds index page

This page is important so that `docs.quantum.ibm.com/api/qiskit-addon-obp` does not 404. It should simply be a flat list of every documented module.

### Flattens submodules

We've gotten prior user complaints that they don't like when docs have a lot of nesting. So, docs.quantum.ibm.com expects all modules to be flattened to the same level in the left ToC. This means that users access submodules via the left ToC, rather than through the parent module's page.

Due to this change, we delete qiskit_addon_obp.utils because it does not actually have any importable interfaces.

If the module list ever gets too big, the qiskit/documentation repository has a mechanism to let you group modules into parent folders (see https://docs.quantum.ibm.com/api/qiskit).

### Sets module metadata & moves structural directives to RST

The script in `qiskit/documentation` expects you to set ``(:mod:`qiskit_addon_obp`)`` at the top of every module page. We need this to extract metadata, which we use to set up the left ToC and `<head>` metadata for the page.

We also now set all the `.. autodoc` directives like `.. autosummary` in RST. Sphinx [doesn't mix well](https://stackoverflow.com/questions/62855567/headings-when-using-autodoc-in-sphinx) when you have some structural elements like headings in RST and some in Python, so it's important to be consistent with where those structural elements go. Putting it in RST is more conventional, and it avoids us having to disable three Ruff lint rules. For modules, most the docs should be in the RST file other than a one-sentence summary of the module.

### Inlines functions and some classes

We've gotten prior user complaints that they don't like when you have to click to a bunch of subpages to see individual functions and classes. So, we've generally adopted the approach of inlining functions and smaller classes. Larger classes and functions can still live on dedicated pages.

The right page ToC makes it so that users can still jump around the page.

This change is subjective. Feel free to push back if you think any particular interface should be a dedicated page!

We set up client-redirects using sphinx-reredirects to avoid 404s. We used this extension well with Qiskit SDK.

## Results

<details><summary>docs.quantum.ibm.com</summary>

<img width="976" alt="Screenshot 2024-10-22 at 3 22 32 PM" src="https://github.com/user-attachments/assets/33bc2458-397f-4541-a7ba-2655b32c8c80">
</details>

<details><summary>GitHub Pages</summary>

<img width="289" alt="Screenshot 2024-10-22 at 3 31 20 PM" src="https://github.com/user-attachments/assets/70bd5d2a-fd05-4fd8-ad0d-efed379688bf">
<img width="625" alt="Screenshot 2024-10-22 at 3 31 23 PM" src="https://github.com/user-attachments/assets/9de98b41-d0b0-44b9-bcce-dbb1844b1057">
</details>

(cherry picked from commit 9675e1b8086f1c03d3d1e1bee236c417f0bb12f4)

# Conflicts:
#	qiskit_addon_obp/utils/__init__.py
#	qiskit_addon_obp/utils/metadata.py
#	qiskit_addon_obp/utils/operations.py
#	qiskit_addon_obp/utils/simplify.py
#	qiskit_addon_obp/utils/truncating.py
#	qiskit_addon_obp/utils/visualization.py