Always show type hints in docs (#31)
Before, we would only show type hints for parameters that had dedicated docstring. There are a lot of parameters that don't have docstring—and sometimes don't need it because it's obvious—so it's best to no matter what show the docstring.

This PR aligns the autodoc templates and config with the other addons like cutting.