How-to guide for noisy circuit simulation using OBP (#85)
* docs: add a guide on how to simulate circuits using OBP

* test: ensure PauliLindbladErrorInstruction._define is covered

* docs: update the noisy expectation value computation

This updates the how-to guide on simulating using OBP based on feedback
I received offline.

The most significant change is the update to the noisy expectation value
computation section, which now resembles a workflow involving the
`NoiseLearner` (and its returned result) more closely.

* fix: ruff formatting

* Apply suggestions from code review

Co-authored-by: Jen Glick <41485571+jenglick@users.noreply.github.com>

* docs: decrease fake noise model effect

Prior to this commit, the random noise model that was being generated
essentially killed the entire signal, resulting in a near-zero
expectation value. This commit reduces the noise model coefficient scale
to ensure that some signal remains, resulting in an expectation value
~85% of the noiseless one.

---------

Co-authored-by: Jen Glick <41485571+jenglick@users.noreply.github.com>